+++ src/expandableCalendar/weekCalendar.js
@@ -1,22 +1,39 @@
 import _ from 'lodash';
 import React, {Component} from 'react';
-import {FlatList} from 'react-native';
+import {FlatList, View, Text} from 'react-native';
 import PropTypes from 'prop-types';
 import XDate from 'xdate';
 
 import styleConstructor from './style';
 import CalendarList from '../calendar-list';
 import Week from '../expandableCalendar/week';
+import asCalendarConsumer from './asCalendarConsumer';
+import {weekDayNames} from '../dateutils';
 
 const commons = require('./commons');
 const UPDATE_SOURCES = commons.UPDATE_SOURCES;
 const NUMBER_OF_PAGES = 2; // must be a positive number
 
+/**
+ * @description: Week calendar component
+ * @example: https://github.com/wix/react-native-calendars/blob/master/example/src/screens/expandableCalendar.js
+ */
 class WeekCalendar extends Component {
+  static displayName = 'WeekCalendar';
+
   static propTypes = {
     ...CalendarList.propTypes,
     // the current date
     current: PropTypes.any,
+    /** whether to have shadow/elevation for the calendar */
+    allowShadow: PropTypes.bool,
+    /** whether to hide the names of the week days */
+    hideDayNames: PropTypes.bool,
+  };
+
+  static defaultProps = {
+    firstDay: 0,
+    allowShadow: true,
   };
 
   constructor(props) {
@@ -33,6 +50,7 @@ class WeekCalendar extends Component {
 
   componentDidUpdate(prevProps) {
     const {updateSource, date} = this.props.context;
+
     if (
       date !== prevProps.context.date &&
       updateSource !== UPDATE_SOURCES.WEEK_SCROLL
@@ -45,6 +63,10 @@ class WeekCalendar extends Component {
     }
   }
 
+  get containerWidth() {
+    return this.props.calendarWidth || commons.screenWidth;
+  }
+
   getDatesArray() {
     const array = [];
     for (let index = -NUMBER_OF_PAGES; index <= NUMBER_OF_PAGES; index++) {
@@ -55,8 +77,8 @@ class WeekCalendar extends Component {
   }
 
   getDate(weekIndex) {
-    const {current, firstDay} = this.props;
-    const d = XDate(current);
+    const {current, context, firstDay} = this.props;
+    const d = XDate(current || context.date);
     // get the first day of the week as date (for the on scroll mark)
     let dayOfTheWeek = d.getDay();
     if (dayOfTheWeek < firstDay && firstDay > 0) {
@@ -71,12 +93,37 @@ class WeekCalendar extends Component {
     return dateString;
   }
 
+  getMarkedDates() {
+    const {context, markedDates} = this.props;
+
+    if (markedDates) {
+      const marked = _.cloneDeep(markedDates);
+
+      if (marked[context.date]) {
+        marked[context.date].selected = true;
+      } else {
+        marked[context.date] = {selected: true};
+      }
+      return marked;
+    }
+    return {[context.date]: {selected: true}};
+  }
+
+  onDayPress = value => {
+    _.invoke(
+      this.props.context,
+      'setDate',
+      value.dateString,
+      UPDATE_SOURCES.DAY_PRESS,
+    );
+  };
+
   onScroll = ({
     nativeEvent: {
       contentOffset: {x},
     },
   }) => {
-    const newPage = Math.round(x / commons.screenWidth);
+    const newPage = Math.round(x / this.containerWidth);
 
     if (this.page !== newPage) {
       const {items} = this.state;
@@ -133,13 +180,16 @@ class WeekCalendar extends Component {
   };
 
   renderItem = ({item}) => {
-    const {calendarWidth, style, ...others} = this.props;
+    const {calendarWidth, style, onDayPress, ...others} = this.props;
+
     return (
       <Week
         {...others}
-        current={item}
         key={item}
-        style={[{width: calendarWidth || commons.screenWidth}, style]}
+        current={item}
+        style={[{width: calendarWidth || this.containerWidth}, style]}
+        markedDates={this.getMarkedDates()}
+        onDayPress={onDayPress || this.onDayPress}
       />
     );
   };
@@ -155,27 +205,51 @@ class WeekCalendar extends Component {
   keyExtractor = (item, index) => index.toString();
 
   render() {
+    const {allowShadow, firstDay, hideDayNames} = this.props;
     const {items} = this.state;
+    let weekDaysNames = weekDayNames(firstDay);
 
     return (
-      <FlatList
-        ref={this.list}
-        data={items}
-        extraData={this.props.current}
-        style={this.style.container}
-        horizontal
-        showsHorizontalScrollIndicator={false}
-        pagingEnabled
-        scrollEnabled
-        renderItem={this.renderItem}
-        keyExtractor={this.keyExtractor}
-        initialScrollIndex={NUMBER_OF_PAGES}
-        getItemLayout={this.getItemLayout}
-        onScroll={this.onScroll}
-        onMomentumScrollEnd={this.onMomentumScrollEnd}
-      />
+      <View
+        style={[
+          allowShadow && this.style.containerShadow,
+          !hideDayNames && {paddingBottom: 6},
+        ]}>
+        {!hideDayNames && (
+          <View style={[this.style.week, {marginTop: 12, marginBottom: -2}]}>
+            {/* {this.props.weekNumbers && <Text allowFontScaling={false} style={this.style.dayHeader}></Text>} */}
+            {weekDaysNames.map((day, idx) => (
+              <Text
+                allowFontScaling={false}
+                key={idx}
+                accessible={false}
+                style={this.style.dayHeader}
+                numberOfLines={1}
+                importantForAccessibility="no">
+                {day}
+              </Text>
+            ))}
+          </View>
+        )}
+        <FlatList
+          ref={this.list}
+          data={items}
+          extraData={this.props.current || this.props.context.date}
+          style={this.style.container}
+          horizontal
+          showsHorizontalScrollIndicator={false}
+          pagingEnabled
+          scrollEnabled
+          renderItem={this.renderItem}
+          keyExtractor={this.keyExtractor}
+          initialScrollIndex={NUMBER_OF_PAGES}
+          getItemLayout={this.getItemLayout}
+          onScroll={this.onScroll}
+          onMomentumScrollEnd={this.onMomentumScrollEnd}
+        />
+      </View>
     );
   }
 }
 
-export default WeekCalendar;
+export default asCalendarConsumer(WeekCalendar);